+++ jstemplate/tests/test_loading.py
@@ -15,26 +15,26 @@ __all__ = ["FindTest", "GetFindersTest"]
 class FindTest(TestCase):
     @property
     def func(self):
-        from mustachejs.loading import find
+        from jstemplate.loading import find
         return find
 
 
-    @patch("mustachejs.loading.finders", [MockFinder("/path/to/a/file.html")])
+    @patch("jstemplate.loading.finders", [MockFinder("/path/to/a/file.html")])
     def test_find(self):
         self.assertEqual(self.func("file"), [("file", "/path/to/a/file.html")])
 
 
     @patch(
-        "mustachejs.loading.finders",
+        "jstemplate.loading.finders",
         [MockFinder(), MockFinder("/path/to/a/file.html")])
     def test_find_fallback(self):
         self.assertEqual(self.func("file"), [("file", "/path/to/a/file.html")])
 
 
-    @patch("mustachejs.loading.finders", [MockFinder()])
+    @patch("jstemplate.loading.finders", [MockFinder()])
     def test_none_found(self):
-        from mustachejs.loading import MustacheJSTemplateNotFound
-        with self.assertRaises(MustacheJSTemplateNotFound):
+        from jstemplate.loading import JSTemplateNotFound
+        with self.assertRaises(JSTemplateNotFound):
             self.func("file")
 
 
@@ -42,11 +42,11 @@ class FindTest(TestCase):
 class GetFindersTest(TestCase):
     @property
     def func(self):
-        from mustachejs.loading import _get_finders
+        from jstemplate.loading import _get_finders
         return _get_finders
 
 
-    @override_settings(MUSTACHEJS_FINDERS=["mustachejs.tests.mockfinders.MockFinder"])
+    @override_settings(JSTEMPLATE_FINDERS=["jstemplate.tests.mockfinders.MockFinder"])
     def test_get_finders(self):
         finders = self.func()
 
@@ -54,14 +54,14 @@ class GetFindersTest(TestCase):
         self.assertIsInstance(finders[0], MockFinder)
 
 
-    @override_settings(MUSTACHEJS_FINDERS=["mustachejs.tests.doesntexist.MockFinder"])
+    @override_settings(JSTEMPLATE_FINDERS=["jstemplate.tests.doesntexist.MockFinder"])
     def test_bad_module(self):
         with self.assertRaises(ImproperlyConfigured):
             self.func()
 
 
     @override_settings(
-        MUSTACHEJS_FINDERS=["mustachejs.tests.mockfinders.DoesntExist"])
+        JSTEMPLATE_FINDERS=["jstemplate.tests.mockfinders.DoesntExist"])
     def test_bad_attribute(self):
         with self.assertRaises(ImproperlyConfigured):
             self.func()